logical: Use VIR_APPEND_ELEMENT instead of VIR_REALLOC_N
Rather than preallocating a set number of elements, then walking through
the extents and adjusting the specific element in place, use the APPEND
macros to handle that chore.

+++ src/conf/storage_conf.h
@@ -50,7 +50,7 @@ struct _virStorageVolSourceExtent {
 typedef struct _virStorageVolSource virStorageVolSource;
 typedef virStorageVolSource *virStorageVolSourcePtr;
 struct _virStorageVolSource {
-    int nextent;
+    size_t nextent;
     virStorageVolSourceExtentPtr extents;
 
     int partType; /* virStorageVolTypeDisk, only used by disk
